DF-11368 compilation fix
Former-commit-id: 910b124ad8bef54e7218ddbc055c2a06eaf52f5f

+++ Sources/Internal/UI/Styles/UIStyleSheetSelectorChain.cpp
@@ -180,6 +180,6 @@ Vector<UIStyleSheetSelector>::const_reverse_iterator UIStyleSheetSelectorChain::
     
 int32 UIStyleSheetSelectorChain::GetSize() const
 {
-    return selectors.size();
+    return (int32)selectors.size();
 }
 }
